requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
joblib==1.4.2
load-dotenv==0.1.0
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.3.0
